fix: coalesce concurrent GetWorkflowMeta loads
singleflight same-ID provider reads in the request cache and move
workflow meta unit tests under the file-length limit.
